Update template variable sytax
The `${}` syntax wasn't working in the `sed` command as expected.  So, replacing `${}` wiith `<>`...at least for now.  Considering eventually replacing `<>` with `{{}}` for future integration with `cookiecutter` for a broader set of templating capabilities.  This could result in a lighter-weight `Makefile` because the code for templating would be deprecated in favor of `cookiecutter`.